requirements.txt
numpy<1.25.0
torch
openai-whisper
faster-whisper
mad-whisper-progress

